feat/refactor: introduce generic object detection skills, modularise from this.
