+++ skills/src/lasr_skills/__init__.py
@@ -1,3 +1,6 @@
+from .detect_objects import DetectObjects
+from .detect_objects_3d import DetectObjects3D
+from .detect_objects_in_area_3d import DetectObjectsInArea3D
 from .detect_people import DetectPeople
 from .detect_people_3d import DetectPeople3D
 from .detect_people_in_area_3d import DetectPeopleInArea3D
