+++ skills/src/lasr_skills/detect_people_in_area_3d.py
@@ -1,32 +1,9 @@
 #!/usr/bin/env python3
-import rospy
-import smach
 
-from lasr_skills import DetectPeople3D
-from lasr_shapely import LasrShapely
+from lasr_skills import DetectObjectsInArea3D
 
-
-class DetectPeopleInArea3D(smach.StateMachine):
-
-
-    class FilterDetections(smach.State):
-
-        def __init__(self):
-            smach.State.__init__(self, outcomes=['succeeded', 'failed'], input_keys=['area_polygon','people_detections_3d'], output_keys=['people_detections_3d'])
-
-            self.shapely = LasrShapely()
-
-        def execute(self, userdata):
-            satisfied_points = self.shapely.are_points_in_polygon_2d(userdata.area_polygon, [[pose[0], pose[1]] for (_, pose) in userdata.people_detections_3d]).inside
-            filtered_detections = [userdata.people_detections_3d[i] for i in range(0, len(userdata.people_detections_3d)) if satisfied_points[i]]
-            userdata.people_detections_3d = filtered_detections
-
-            return 'succeeded'
+class DetectPeopleInArea3D(DetectObjectsInArea3D):
 
     def __init__(self):
-        smach.StateMachine.__init__(self, outcomes=['succeeded', 'failed'], input_keys=['pcl_msg', 'area_polygon'], output_keys=['people_detections_3d'])
-
-        with self:
-
-            smach.StateMachine.add('DETECT_PEOPLE_3D', DetectPeople3D(), transitions={'succeeded': 'FILTER_DETECTIONS', 'failed' : 'failed'})
-            smach.StateMachine.add('FILTER_DETECTIONS', self.FilterDetections(), transitions={'succeeded' : 'succeeded', 'failed' : 'failed'})
+        self.userdata.filter = ["person"]
+        super().__init__()
