+++ skills/src/lasr_skills/wait_for_person_in_area.py
@@ -20,16 +20,16 @@ class WaitForPersonInArea(smach.StateMachine):
     class CheckForPerson(smach.State):
 
         def __init__(self):
-            smach.State.__init__(self, outcomes=['done', 'not_done'], input_keys=['people_detections_3d'])
+            smach.State.__init__(self, outcomes=['done', 'not_done'], input_keys=['detections_3d'])
 
         def execute(self, userdata):
-            if len(userdata.people_detections_3d):
+            if len(userdata.detections_3d):
                 return 'done'
             else:
                 return 'not_done'
             
     def __init__(self):
-        smach.StateMachine.__init__(self, outcomes=['succeeded', 'failed'], input_keys=['area_polygon'], output_keys=['people_detections_3d'])
+        smach.StateMachine.__init__(self, outcomes=['succeeded', 'failed'], input_keys=['area_polygon'], output_keys=['detections_3d'])
 
 
         with self:
